% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap_year.R
\name{leap_year}
\alias{leap_year}
\title{LeapYear}
\usage{
leap_year(year)
}
\arguments{
\item{year}{Enter one year between 1804 & 2400.  This must be a vector of length 1.}
}
\value{
The function will return whether or not the year entered as an argument is a leap year.
}
\description{
This package has on function that declares whether or not the year entered as an argument is a leap year.
}
\note{
The package contains a vector of leap year titled 'leap_years' that the leap_year() function uses.
}
\examples{
\dontrun{
leap_year(1992)
}

}
