% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap_years.R
\docType{data}
\name{leap_years}
\alias{leap_years}
\title{This is leap year data to be included in the package}
\format{vector of length 146}
\usage{
data(leap_years)
}
\description{
This is leap year data to be included in the package
}
\keyword{data}
